# openapi.yaml
openapi: 3.1.0
info:
  title: StockIt REST API
  version: 1.0.0
  description: |
    Machine-readable integration contract for coding agents building warehouse applications on StockIt.
    Discover accessible tables and their fields at runtime before using generic record operations.
    The REST API and MCP tools are intentionally aligned: table catalog/schema discovery,
    record list/get/create/update/delete, and shared session authentication. MCP transport is
    documented in README.md and StokIt_Specification.md rather than OpenAPI.
    StockIt owns one workflow: purchase order approval. Orders are priced from
    their lines, routed through value-based approval rules, and converted to purchase orders
    (see the Approvals tag). Client applications still own import, calculation, forecasting,
    consistency-check, and UI workflows.

    List endpoints filter server-side: `filter.<column>` for equality, `from.<column>` and
    `to.<column>` for inclusive ranges, and `q` for a substring search across text columns.

    Cash-planning tables include bank_accounts, designation_codes, financial_obligations,
    and bank_transactions. Financial obligations represent expected payable or receivable
    payments, including multiple purchase-order installments and one sales-order payment.
    Money fields use integer minor units with an accompanying currency; bank transaction
    amounts are signed (positive inflow, negative outflow). Bank balances are derived from
    bank transactions.
servers:
  - url: http://127.0.0.1:8080
    description: Default HTTP listener
  - url: https://127.0.0.1:8443
    description: Default HTTPS listener with startup-generated self-signed certificate
tags:
  - name: Auth
  - name: Metadata
  - name: Records
  - name: Approvals
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate a StockIt user and return a bearer token
      description: |
        For trusted external app backends, submit the user's credentials server-to-server over HTTPS.
        Keep the returned token only in the backend session, then use it in Authorization: Bearer
        requests so StockIt preserves the authenticated user's role and record attribution. Do not
        expose or persist the token in browser code. This trial flow is not delegated SSO.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          headers:
            Set-Cookie:
              description: Session cookie for browser-style clients
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Session limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Invalidate the current session
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/me:
    get:
      tags: [Auth]
      summary: Get the current authenticated principal
      responses:
        '200':
          description: Current session principal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/tables:
    get:
      tags: [Metadata]
      summary: List accessible top-level tables
      responses:
        '200':
          description: Visible tables for the current role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableCatalogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/tables/{table}/schema:
    get:
      tags: [Metadata]
      summary: Describe a table schema
      parameters:
        - $ref: '#/components/parameters/TableName'
      responses:
        '200':
          description: Table schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableSchemaResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tables/{table}:
    get:
      tags: [Records]
      summary: List records from a table
      parameters:
        - $ref: '#/components/parameters/TableName'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Desc'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/ParentID'
        - $ref: '#/components/parameters/ParentField'
        - $ref: '#/components/parameters/ColumnFilter'
        - $ref: '#/components/parameters/ColumnFrom'
        - $ref: '#/components/parameters/ColumnTo'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Table rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRowsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Records]
      summary: Create a record in a table
      parameters:
        - $ref: '#/components/parameters/TableName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordWriteRequest'
      responses:
        '201':
          description: Created record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tables/{table}/{id}:
    get:
      tags: [Records]
      summary: Fetch one record by primary key
      parameters:
        - $ref: '#/components/parameters/TableName'
        - $ref: '#/components/parameters/RecordID'
      responses:
        '200':
          description: Table row
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Records]
      summary: Update one record by primary key
      parameters:
        - $ref: '#/components/parameters/TableName'
        - $ref: '#/components/parameters/RecordID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordWriteRequest'
      responses:
        '200':
          description: Updated row
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Records]
      summary: Delete one record by primary key
      parameters:
        - $ref: '#/components/parameters/TableName'
        - $ref: '#/components/parameters/RecordID'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Deleting the last admin user is blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/{table}/import:
    post:
      tags: [Records]
      summary: Bulk-import rows from a CSV file
      description: |
        Accepts a CSV document (header row required) and inserts each data row
        through the validated application layer, applying the same per-field
        parsing and business-rule checks as individual create calls. Partial
        failure returns an error and the rows already inserted remain
        committed.
      parameters:
        - $ref: '#/components/parameters/TableName'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [csv_file]
              properties:
                csv_file:
                  type: string
                  format: binary
                  description: CSV document. First row is the header; column names match table field columns or labels.
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableImportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'


  /api/approvals/{id}/decide:
    post:
      tags: [Approvals]
      summary: Approve or reject one pending approval step
      description: |
        Steps are decided in order and only by a user whose role matches the
        step's `apv_role`. A rejection sets the purchase order to `rejected`;
        approving the last pending step sets it to `approved`. Both moves are
        appended to `po_status_history`.
      parameters:
        - $ref: '#/components/parameters/RecordID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [approved, rejected]
                note:
                  type: string
      responses:
        '200':
          description: Decision result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDecision'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/purchase_orders/{id}/submit:
    post:
      tags: [Approvals]
      summary: Submit a draft purchase order for approval
      description: |
        Prices the purchase order from its lines, stores the total in
        `por_total_minor`, and creates one pending approval step per active
        approval rule with `apr_source_type = purchase_order` whose
        `apr_min_amount_minor` the total reaches. An order that matches no rule
        is approved outright. Only a draft order with at least one line can be
        submitted; the status moves to `pending_approval`, then `approved` or
        `rejected`. Every one of those moves is appended to `po_status_history`.
      parameters:
        - $ref: '#/components/parameters/RecordID'
      responses:
        '200':
          description: Submission result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalSubmission'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/purchase_orders/{id}/status:
    post:
      tags: [Approvals]
      summary: Change a purchase order's status or payment tag with a note
      description: |
        Records the change in `po_status_history` together with the acting user,
        the timestamp and the note. Transitions are unrestricted by design: the
        history is the control, not a state machine. Send only the field you
        want to change; an empty value leaves that field alone. Status changes
        made through `PUT /api/tables/purchase_orders/{id}` are recorded too,
        but with no note.
      parameters:
        - $ref: '#/components/parameters/RecordID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                por_status:
                  type: string
                  enum: [draft, pending_approval, approved, rejected, issued, confirmed, partially_received, received, invoiced, closed, on_hold, pending_revision, cancelled]
                por_payment_status:
                  type: string
                  enum: [unpaid, partially_paid, paid, refunded]
                note:
                  type: string
      responses:
        '200':
          description: Updated purchase order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrderStatusChange'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque session token
    sessionCookie:
      type: apiKey
      in: cookie
      name: stockit_session
  parameters:
    TableName:
      name: table
      in: path
      required: true
      schema:
        type: string
        enum:
          - users
          - customers
          - suppliers
          - locations
          - items
          - boms
          - bom_components
          - purchase_orders
          - po_components
          - quotes
          - quote_components
          - sales_orders
          - sales_order_components
          - manufacturing_orders
          - mfo_components
          - invoices
          - invoice_components
          - adjustments
          - adjustment_components
          - stock_moves
          - bank_accounts
          - designation_codes
          - financial_obligations
          - bank_transactions
          - approval_rules
          - approvals
          - po_status_history
    RecordID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Table primary key value. Current built-in tables use integer identifiers.
    ColumnFilter:
      name: filter.{column}
      in: query
      description: >-
        Exact-match filter on a declared column, for example
        `filter.por_status=draft`. Unknown columns are rejected with 400.
      schema:
        type: string
    ColumnFrom:
      name: from.{column}
      in: query
      description: >-
        Inclusive lower bound on a declared column, for example
        `from.por_doc_date=2026-01-01`.
      schema:
        type: string
    ColumnTo:
      name: to.{column}
      in: query
      description: >-
        Inclusive upper bound on a declared column, for example
        `to.por_doc_date=2026-12-31`.
      schema:
        type: string
    Search:
      name: q
      in: query
      description: Case-insensitive substring matched against the table's text columns.
      schema:
        type: string
    Sort:
      name: sort
      in: query
      schema:
        type: string
    Desc:
      name: desc
      in: query
      schema:
        type: boolean
        default: false
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 30
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    ParentID:
      name: parent_id
      in: query
      description: |
        Restrict the list to rows whose parent foreign key equals this id.
        Only valid for subtables (e.g. bom_components, po_components). If
        omitted, the full table is returned.
      schema:
        type: string
    ParentField:
      name: parent_field
      in: query
      description: |
        Optional parent column name. Defaults to the subtable's declared
        parent field and must match it when provided; this prevents filtering
        by arbitrary columns.
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated role cannot access this table or operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    PurchaseOrderStatusChange:
      type: object
      required: [table, purchase_order]
      properties:
        table:
          type: string
          enum: [purchase_orders]
        purchase_order:
          type: object
          additionalProperties: true
    ApprovalSubmission:
      type: object
      required: [source_type, source_id, status, total_minor, approvals]
      properties:
        source_type:
          type: string
          enum: [purchase_order]
        source_id:
          type: integer
        status:
          type: string
          enum: [pending_approval, approved]
        total_minor:
          type: integer
          description: Purchase order value in integer minor units, priced from its lines.
        currency:
          type: string
        approvals:
          type: array
          description: Pending approval steps, in decision order.
          items:
            type: object
            additionalProperties: true
    ApprovalDecision:
      type: object
      required: [approval, source_type, source_id, status, remaining_approvals]
      properties:
        approval:
          type: object
          additionalProperties: true
        source_type:
          type: string
          enum: [purchase_order]
        source_id:
          type: integer
        status:
          type: string
          description: Lifecycle status of the purchase order after the decision.
          enum: [pending_approval, approved, rejected]
        remaining_approvals:
          type: integer
    TableImportResponse:
      type: object
      additionalProperties: false
      required: [table, imported]
      properties:
        table:
          type: string
        imported:
          type: integer
          minimum: 0